+++ Assets/Scripts/Character/CharacterSearch.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f822ab56d65af3c42b7f7664b4c9529a
+guid: 083cec9a5b2d93d4eb933b6c35d0212a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
