--- Assets/Scripts/Character/CharacterSearchVer2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 083cec9a5b2d93d4eb933b6c35d0212a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
